# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthcare Insurance Claims Automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ClaimStatus = (typeof ClaimStatus)[keyof typeof ClaimStatus];

export const ClaimStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  DENIED: "DENIED",
  SUBMITTED: "SUBMITTED",
  PROCESSING: "PROCESSING",
  RESUBMITTED: "RESUBMITTED",
  PAID: "PAID",
} as const;

export type AgentStepStatus =
  (typeof AgentStepStatus)[keyof typeof AgentStepStatus];

export const AgentStepStatus = {
  pending: "pending",
  running: "running",
  done: "done",
  failed: "failed",
  skipped: "skipped",
} as const;

export interface AgentStep {
  agent: string;
  status: AgentStepStatus;
  startedAt: string | null;
  completedAt: string | null;
  output: string | null;
}

export type ClaimEventData = { [key: string]: unknown } | null;

export interface ClaimEvent {
  id: string;
  claimId: string;
  agentName: string;
  eventType: string;
  message: string;
  data: ClaimEventData;
  createdAt: string;
}

export interface Claim {
  id: string;
  patientId: string;
  patientName: string;
  insurerId: string;
  insurerName: string;
  status: ClaimStatus;
  riskScore: number;
  issuesDetected: string[];
  actionsTaken: string[];
  icdCodes: string[];
  cptCodes: string[];
  totalAmount: number;
  approvedAmount: number | null;
  paymentStatus: string | null;
  denialReason: string | null;
  submittedAt: string | null;
  resolvedAt: string | null;
  createdAt: string;
  agentSteps: AgentStep[];
}

export interface ClaimListResponse {
  claims: Claim[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateClaimRequest {
  patientId: string;
  doctorNotes: string;
  insurerId: string;
  totalAmount: number;
}

export type PatientEligibilityStatus =
  (typeof PatientEligibilityStatus)[keyof typeof PatientEligibilityStatus];

export const PatientEligibilityStatus = {
  ELIGIBLE: "ELIGIBLE",
  INELIGIBLE: "INELIGIBLE",
  PENDING: "PENDING",
} as const;

export interface Patient {
  id: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  memberId: string;
  groupNumber: string;
  insurerId: string;
  insurerName: string;
  eligibilityStatus: PatientEligibilityStatus;
  createdAt: string;
}

export interface PatientListResponse {
  patients: Patient[];
  total: number;
}

export interface CreatePatientRequest {
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  memberId: string;
  groupNumber: string;
  insurerId: string;
}

export interface RunAgentRequest {
  claimId: string;
  agentName: string | null;
}

export interface AgentPipelineResult {
  claimId: string;
  status: ClaimStatus;
  riskScore: number;
  issuesDetected: string[];
  actionsTaken: string[];
  paymentStatus: string | null;
}

export interface DashboardSummary {
  totalClaims: number;
  pendingClaims: number;
  approvedClaims: number;
  deniedClaims: number;
  paidClaims: number;
  totalRevenue: number;
  pendingRevenue: number;
  denialRate: number;
  avgRiskScore: number;
  avgProcessingTimeHours: number;
}

export interface RecentActivityItem {
  id: string;
  claimId: string;
  patientName: string;
  agentName: string;
  message: string;
  status: ClaimStatus;
  createdAt: string;
}

export interface RecentActivityResponse {
  activities: RecentActivityItem[];
}

export interface DenialReasonItem {
  reason: string;
  count: number;
  percentage: number;
}

export interface DenialAnalysisResponse {
  denialReasons: DenialReasonItem[];
  totalDenied: number;
  resubmissionSuccessRate: number;
}

export interface ClaimEventsResponse {
  events: ClaimEvent[];
}

export type RevenueStatsResponseMonthlyBreakdownItem = {
  month: string;
  expected: number;
  actual: number;
};

export interface RevenueStatsResponse {
  expectedRevenue: number;
  actualRevenue: number;
  underpaymentAmount: number;
  discrepancyCount: number;
  monthlyBreakdown: RevenueStatsResponseMonthlyBreakdownItem[];
}

export type ListClaimsParams = {
  status?: ListClaimsStatus;
  page?: number;
  limit?: number;
};

export type ListClaimsStatus =
  (typeof ListClaimsStatus)[keyof typeof ListClaimsStatus];

export const ListClaimsStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  DENIED: "DENIED",
  SUBMITTED: "SUBMITTED",
  PROCESSING: "PROCESSING",
  RESUBMITTED: "RESUBMITTED",
  PAID: "PAID",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};