# lib/api-zod/src/generated/types/claimStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthcare Insurance Claims Automation API
 * OpenAPI spec version: 0.1.0
 */

export type ClaimStatus = (typeof ClaimStatus)[keyof typeof ClaimStatus];

export const ClaimStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  DENIED: "DENIED",
  SUBMITTED: "SUBMITTED",
  PROCESSING: "PROCESSING",
  RESUBMITTED: "RESUBMITTED",
  PAID: "PAID",
} as const;