# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthcare Insurance Claims Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalClaims: number;
  pendingClaims: number;
  approvedClaims: number;
  deniedClaims: number;
  paidClaims: number;
  totalRevenue: number;
  pendingRevenue: number;
  denialRate: number;
  avgRiskScore: number;
  avgProcessingTimeHours: number;
}