# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthcare Insurance Claims Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientEligibilityStatus } from "./patientEligibilityStatus";

export interface Patient {
  id: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  memberId: string;
  groupNumber: string;
  insurerId: string;
  insurerName: string;
  eligibilityStatus: PatientEligibilityStatus;
  createdAt: string;
}